🪄 #21 surrounded palaces
