+++ src/astro/FunctionalAstrolabe.ts
@@ -1,10 +1,16 @@
 import { getHeavenlyStemAndEarthlyBranchBySolarDate, normalizeSolarDateStr, solar2lunar } from '../calendar';
 import { EARTHLY_BRANCHES } from '../data';
-import { Astrolabe, Horoscope } from '../data/types';
+import { Astrolabe, Horoscope, SurroundedPalaces } from '../data/types';
 import { EarthlyBranchKey, EarthlyBranchName, HeavenlyStemName, kot, PalaceName, StarName } from '../i18n';
 import { getHoroscopeStar } from '../star';
 import { fixEarthlyBranchIndex, fixIndex, getMutagensByHeavenlyStem, timeToIndex } from '../utils';
-import { getPalace, isSurroundedByStars } from './analyzer';
+import {
+  getPalace,
+  getSurroundedPalaces,
+  isSurroundedByOneOfStars,
+  isSurroundedByStars,
+  notSurroundedByStars,
+} from './analyzer';
 import { IFunctionalPalace } from './FunctionalPalace';
 import { getPalaceNames } from './palace';
 
@@ -168,6 +174,16 @@ export interface IFunctionalAstrolabe extends Astrolabe {
   palace: (indexOrName: number | PalaceName) => IFunctionalPalace | undefined;
 
   /**
+   * 获取三方四正宫位，所谓三方四正就是传入的目标宫位，以及其对宫，财帛位和官禄位，总共四个宫位
+   *
+   * @version v1.1.0
+   *
+   * @param indexOrName 宫位索引或者宫位名称
+   * @returns 三方四正宫位
+   */
+  surroundedPalaces: (indexOrName: number | PalaceName) => SurroundedPalaces;
+
+  /**
    * 判断某一个宫位三方四正是否包含目标星耀，必须要全部包含才会返回true
    *
    * @param indexOrName 宫位索引或者宫位名称
@@ -175,6 +191,28 @@ export interface IFunctionalAstrolabe extends Astrolabe {
    * @returns true | false
    */
   isSurrounded: (indexOrName: number | PalaceName, stars: StarName[]) => boolean;
+
+  /**
+   * 判断三方四正内是否有传入星耀的其中一个，只要命中一个就会返回true
+   *
+   * @version v1.1.0
+   *
+   * @param indexOrName 宫位索引或者宫位名称
+   * @param stars 星耀名称数组
+   * @returns true | false
+   */
+  isSurroundedOneOf: (indexOrName: number | PalaceName, stars: StarName[]) => boolean;
+
+  /**
+   * 判断某一个宫位三方四正是否不含目标星耀，必须要全部都不在三方四正内含才会返回true
+   *
+   * @version v1.1.0
+   *
+   * @param indexOrName 宫位索引或者宫位名称
+   * @param stars 星耀名称数组
+   * @returns true | false
+   */
+  notSurrounded: (indexOrName: number | PalaceName, stars: StarName[]) => boolean;
 }
 
 export default class FunctionalAstrolabe implements IFunctionalAstrolabe {
@@ -215,6 +253,14 @@ export default class FunctionalAstrolabe implements IFunctionalAstrolabe {
 
   palace = (indexOrName: number | PalaceName): IFunctionalPalace | undefined => getPalace(this, indexOrName);
 
+  surroundedPalaces = (indexOrName: number | PalaceName): SurroundedPalaces => getSurroundedPalaces(this, indexOrName);
+
   isSurrounded = (indexOrName: number | PalaceName, stars: StarName[]): boolean =>
     isSurroundedByStars(this, indexOrName, stars);
+
+  isSurroundedOneOf = (indexOrName: number | PalaceName, stars: StarName[]): boolean =>
+    isSurroundedByOneOfStars(this, indexOrName, stars);
+
+  notSurrounded = (indexOrName: number | PalaceName, stars: StarName[]): boolean =>
+    notSurroundedByStars(this, indexOrName, stars);
 }
