+++ src/data/types/index.ts
@@ -1,0 +1,5 @@
+export * from './general';
+export * from './calendar';
+export * from './star';
+export * from './astro';
+export * from './palace';
